backend/pyproject.toml
[project]
name = "clargate-backend"
version = "0.1.0"
description = "Clargate IRB Platform - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "boto3>=1.35.0",
    "anthropic>=0.40.0",
    "openai>=1.60.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
